test!(virt_text): fix, improve and rename util func
Previous GetVirtualTexts() has got every chunks in the same line as if
a set of different text.

Fixing the problem above, this commit updates the function so that it
returns a dictionary each value of which represents one plain text by
a line number.

Moreover, rename the function more informative.